Hero icons now safe for fingerprinting (prod env)
Hero icon paths were previously hard-coded, which doesn't work with the
fingerprinting done by rails asset pipeline in production environment.
Now a map of hero names to icon paths is included in the bottom of the
show.html.erb page, and is subsequently used to locate hero icons for
use on the hero positions map.